requirements.txt
flask==3.1.0
pdfplumber==0.11.4
Pillow==11.1.0
gunicorn==23.0.0
